Add min. to handle static export.
The only thing stopping this from being independant of an active server running it, is as far as I can tell the api routes.

We try to make the paths to the functions that the FE calls less hard and basically allow for whoever is consuming the FE to choose their own experiement endpoint.

This isn't as nice as being able to use the build in api routes but it does allow us to host it statically.

If we were to abstract the api routes a bit we would also be able to make use of the logic in them and have them run in their own seperate environment such as lambda or any other http server for that matter.